requirements.txt
fastapi
cx_Oracle
google-generativeai
langchain-google-genai
apscheduler
pandas